refactor course-list and move course-score-calculation from course into course-list
