+++ src/main/webapp/app/course-list/course-list.component.html
