+++ src/main/webapp/app/course-list/course-score-calculation.component.html
