+++ src/main/webapp/app/course-list/exercise-list/exercise-list.component.html
