+++ src/main/webapp/app/course-list/index.ts
@@ -1,0 +1,7 @@
+export * from './course-list.component';
+export * from './course-list.module';
+export * from './course-list.route';
+export * from './course-score-calculation.component';
+export * from './course-score-calculation.service';
+export * from './exercise-list/exercise-list.component';
+export * from './exercise-list/participation-data-provider';
