--- src/main/webapp/app/courses/courses.route.ts
@@ -1,25 +1,0 @@
-import { Routes } from '@angular/router';
-
-import { UserRouteAccessService } from '../core';
-import { CoursesComponent } from './courses.component';
-
-export const coursesRoute: Routes = [
-    {
-        path: 'courses',
-        component: CoursesComponent,
-        data: {
-            authorities: ['ROLE_USER'],
-            pageTitle: 'arTeMiSApp.course.home.title'
-        },
-        canActivate: [UserRouteAccessService]
-    },
-    {
-        path: 'courses/:courseId/exercise/:exerciseId',
-        component: CoursesComponent,
-        data: {
-            authorities: ['ROLE_USER'],
-            pageTitle: 'arTeMiSApp.course.home.title'
-        },
-        canActivate: [UserRouteAccessService]
-    }
-];
