--- src/main/webapp/app/courses/index.ts
@@ -1,5 +1,0 @@
-export * from './courses.component';
-export * from './courses.module';
-export * from './courses.route';
-export * from './exercises/exercise-list.component';
-export * from './exercises/participation-data-provider';
