+++ src/main/webapp/app/entities/course/course.module.ts
@@ -8,13 +8,11 @@ import {
     CourseDeletePopupComponent,
     CourseDetailComponent,
     CourseExerciseService,
-    CourseScoreCalculationComponent,
     CourseService,
     CourseUpdateComponent,
     coursePopupRoute,
     courseRoute
 } from './';
-import { CourseScoreCalculationService } from './course-score-calculation.service';
 import { FormDateTimePickerModule } from '../../shared/dateTimePicker/date-time-picker.module';
 
 const ENTITY_STATES = [...courseRoute, ...coursePopupRoute];
@@ -26,17 +24,15 @@ const ENTITY_STATES = [...courseRoute, ...coursePopupRoute];
         CourseDetailComponent,
         CourseDeleteDialogComponent,
         CourseUpdateComponent,
-        CourseDeletePopupComponent,
-        CourseScoreCalculationComponent
+        CourseDeletePopupComponent
     ],
     entryComponents: [
         CourseComponent,
         CourseUpdateComponent,
         CourseDeleteDialogComponent,
-        CourseDeletePopupComponent,
-        CourseScoreCalculationComponent
+        CourseDeletePopupComponent
     ],
-    providers: [CourseService, CourseExerciseService, CourseScoreCalculationService],
+    providers: [CourseService, CourseExerciseService],
     schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class ArTEMiSCourseModule {}
