+++ src/main/webapp/app/entities/course/course.route.ts
@@ -10,7 +10,6 @@ import { CourseComponent } from './course.component';
 import { CourseDetailComponent } from './course-detail.component';
 import { CourseUpdateComponent } from './course-update.component';
 import { CourseDeletePopupComponent } from './course-delete-dialog.component';
-import { CourseScoreCalculationComponent } from './course-score-calculation.component';
 
 @Injectable({ providedIn: 'root' })
 export class CourseResolve implements Resolve<Course> {
@@ -70,15 +69,6 @@ export const courseRoute: Routes = [
         },
         data: {
             authorities: ['ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
-            pageTitle: 'arTeMiSApp.course.home.title'
-        },
-        canActivate: [UserRouteAccessService]
-    },
-    {
-        path: 'course/:id/score-calculation',
-        component: CourseScoreCalculationComponent,
-        data: {
-            authorities: ['ROLE_USER', 'ROLE_TA', 'ROLE_INSTRUCTOR', 'ROLE_ADMIN'],
             pageTitle: 'arTeMiSApp.course.home.title'
         },
         canActivate: [UserRouteAccessService]
