+++ src/main/webapp/app/entities/course/index.ts
@@ -1,9 +1,9 @@
 export * from './course.model';
 export * from './course.service';
 export * from './course-update.component';
-export * from './course-score-calculation.service';
+export * from '../../course-list/course-score-calculation.service';
 export * from './course-delete-dialog.component';
-export * from './course-score-calculation.component';
+export * from '../../course-list/course-score-calculation.component';
 export * from './course-detail.component';
 export * from './course.component';
 export * from './course.route';
